rename file plus add  test case
